third_party/WebKit/LayoutTests/inspector/console/console-dir-es6-expected.txt

CONSOLE MESSAGE: line 15: [object Object]
CONSOLE MESSAGE: line 18: Symbol()
CONSOLE MESSAGE: line 22: [object Map]
CONSOLE MESSAGE: line 22: [object WeakMap]
CONSOLE MESSAGE: line 26: [object Set]
CONSOLE MESSAGE: line 26: [object WeakSet]
CONSOLE MESSAGE: line 34: [object Set]
CONSOLE MESSAGE: line 37: [object WeakMap]
CONSOLE MESSAGE: line 45: [object Map Iterator]
CONSOLE MESSAGE: line 46: [object Map Iterator]
CONSOLE MESSAGE: line 47: [object Map Iterator]
CONSOLE MESSAGE: line 45: [object Set Iterator]
CONSOLE MESSAGE: line 46: [object Set Iterator]
CONSOLE MESSAGE: line 47: [object Set Iterator]
CONSOLE MESSAGE: line 50: [object Object]
Tests that console logging dumps proper messages.

console-dir-es6.html:15 Object
    a: 1
    Symbol(): 2
    Symbol(Symbol.iterator): Symbol(foo)
    Symbol(a): 3
    Symbol(a): Symbol(Symbol.iterator)
    __proto__: Object
console-dir-es6.html:18 Symbol()
console-dir-es6.html:22 Map
    size: (...)
    __proto__: Map
    [[Entries]]: Array[1]
        0: {Object => Object}
        length: 1
console-dir-es6.html:22 WeakMap
    __proto__: WeakMap
    [[Entries]]: Array[1]
        0: {Object => Object}
        length: 1
console-dir-es6.html:26 Set
    size: (...)
    __proto__: Set
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:26 WeakSet
    __proto__: WeakSet
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:34 Set
    size: (...)
    __proto__: Set
    [[Entries]]: Array[1]
        0: Set
        length: 1
console-dir-es6.html:37 WeakMap
    __proto__: WeakMap
    [[Entries]]: Array[0]
        length: 0
console-dir-es6.html:45 MapIterator
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "keys"
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:46 MapIterator
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:47 MapIterator
    __proto__: Map Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "entries"
    [[Entries]]: Array[1]
        0: Array[2]
        length: 1
console-dir-es6.html:45 SetIterator
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:46 SetIterator
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "values"
    [[Entries]]: Array[1]
        0: Object
        length: 1
console-dir-es6.html:47 SetIterator
    __proto__: Set Iterator
    [[IteratorHasMore]]: true
    [[IteratorIndex]]: 0
    [[IteratorKind]]: "entries"
    [[Entries]]: Array[1]
        0: Array[2]
        length: 1
console-dir-es6.html:50 Object
    a: ()
    __proto__: Object
